20260212/requirements_video.txt
pandas>=2.0.0
python-dotenv>=1.0.0
langfuse>=2.0.0
requests>=2.31.0
